Add fCreateInvertedPendulumStateSpaceModel
Provides a function which creates a state-space model of inverted
pendulum. Mass and coefficient of friction for cart are constant. The
function parameters are mass of the pendulum (m), length to pendulum
center of mass (l), and mass moment of inertia of the pendulum (I). The
function calculates A, B, C, and D matrices and returns ss model.
Returned model can be used for further simulation.